Add files via upload
This is a project I have finished working, this program is about a familiar board game called tic tac toe game, where two players choose one of two the markers "O" or "X" and take a turn to put their marker in the board. The player who can get their markers forming a straight line (vertical, horizontal or diagonal) win.  This game is written in python for Jupyter notebook.